Rename `api.py` to `init_db.py` :arrows_counterclockwise:
Rename `api.py` to `init_db.py` to indicate it's primary purpose which is to initialize the `bandit` and `parameters` tables in the database and set up the connection between the UI and the api/DB.
