+++ ui/ui_survey.py
@@ -1,8 +1,5 @@
-from api import add_choices, get_choices, top_param
-import os
 from pathlib import Path
 from shiny import ui
-from htmltools import head_content
 
 """
 This script lays out the UI for the cartoon comparison page. It also does a
